serverless/update-github-web-files/requirements.txt
PyGithub
boto3
requests
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
